+++ lib/decorator.py
@@ -1,0 +1,3 @@
+def shell_command(func):
+    func._is_shell_command = True
+    return func